Move TestLifecycleScopeProvider to new artifact
